+++ .gitignore
@@ -1,0 +1,4 @@
+bin
+lib/python3.9
+pyvenv*
+.vscode
