+++ includes/builtin.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   builtin.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: woonchoi <woonchoi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jasong <jasong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/09 17:08:33 by jasong            #+#    #+#             */
-/*   Updated: 2022/06/12 19:11:29 by woonchoi         ###   ########.fr       */
+/*   Updated: 2022/06/13 15:37:46 by jasong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,10 +23,18 @@ int			builtin_export(char **argv, t_info *info);
 int			builtin_pwd(void);
 int			builtin_unset(char **argv, t_env_list *env_head);
 
-t_env_list	*env_key_location(t_env_list *env_head, char *key);
+/* BUILTIN ERROR */
 void		ft_error(char *cmd, char *cmd_arg, char *errmsg);
 void		ft_s_quote_error(char *cmd, char *cmd_arg, char *errmsg);
+
+/* UNSET FUNCTION */
 int			check_avaliable_key(char *key);
 void 		free_split_value(char **s_val);
+
+/* BUILTIN UTILS */
+t_env_list	*new_env_list(char *argv);
+void		env_add_back(t_env_list **env_list, t_env_list *new);
+void		update_env_val(t_env_list *key_loc, char *sep, int del);
+t_env_list	*env_key_location(t_env_list *env_head, char *key);
 
 #endif
