+++ srcs/builtin/builtin_export.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   builtin_export.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: woonchoi <woonchoi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jasong <jasong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/09 17:10:41 by jasong            #+#    #+#             */
-/*   Updated: 2022/06/12 19:11:29 by woonchoi         ###   ########.fr       */
+/*   Updated: 2022/06/13 15:42:28 by jasong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,50 +14,6 @@
 #define NO_DEL 0
 #define ADD_VAL 1
 #define ADD_ENV 2
-
-t_env_list	*new_env_list(char *argv)
-{
-	t_env_list	*new;
-	char		*sep;
-
-	new = (t_env_list *)malloc(sizeof(t_env_list));
-	if (!new)
-		return (NULL);
-	sep = ft_strchr(argv, '=');
-	new->key = ft_strndup(argv, sep - argv);
-	if (!new->key)
-	{
-		free(new);
-		return (NULL);
-	}
-	new->value = ft_strdup(sep + 1);
-	if (!new->value || !check_avaliable_key(new->key))
-	{
-		if (new->value)
-			free(new->value);
-		free(new->key);
-		free(new);
-		return (NULL);
-	}
-	new->split_value = ft_split(sep + 1, ':');
-	new->next = NULL;
-	return (new);
-}
-
-void	env_add_back(t_env_list **env_list, t_env_list *new)
-{
-	t_env_list	*cur;
-
-	if (!*env_list)
-		*env_list = new;
-	else
-	{
-		cur = *env_list;
-		while (cur->next)
-			cur = cur->next;
-		cur->next = new;
-	}
-}
 
 void	print_export_env(t_env_list *env_head)
 {
@@ -80,7 +36,6 @@ int	add_env(char *argv, int del, t_env_list **env_head)
 {
 	char		*sep;
 	char		*key;
-	char		*new_value;
 	t_env_list	*key_loc;
 	t_env_list	*new_env;
 
@@ -93,18 +48,7 @@ int	add_env(char *argv, int del, t_env_list **env_head)
 	key_loc = env_key_location(*env_head, key);
 	free(key);
 	if (key_loc)
-	{
-		if (del == '+')
-			new_value = ft_strdup(sep + 2);
-		else
-		{
-			new_value = ft_strdup(sep + 1);
-			free(key_loc->value);
-			free_split_value(key_loc->split_value);
-			key_loc->value = new_value;
-			key_loc->split_value = ft_split(new_env, ':');
-		}
-	}
+		update_env_val(key_loc, sep, del);
 	else
 	{
 		new_env = new_env_list(argv);
@@ -166,20 +110,14 @@ int	builtin_export(char *argv[], t_info *info)
 	{
 		ret = 1;
 		delimeter = check_delimeter(argv[i]);
-		if (delimeter == NO_DEL)
-		{
-			if (check_avaliable_key(argv[i]))
-				ret = 0;
-		}
+		if (delimeter == NO_DEL && check_avaliable_key(argv[i]))
+			ret = 0;
 		else if (delimeter == ADD_VAL)
 			ret = add_value(argv[i], info->env_head);
 		else if (delimeter == ADD_ENV)
 			ret = add_env(argv[i], '=', &info->env_head);
 		if (ret)
-		{
 			ft_s_quote_error("export", argv[i], "not a valid identifier");
-			return (1);
-		}
 	}
-	return (0);
+	return (ret);
 }
