+++ srcs/builtin/builtin_pwd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   builtin_pwd.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: woonchoi <woonchoi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jasong <jasong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/08 19:56:21 by jasong            #+#    #+#             */
-/*   Updated: 2022/06/13 15:10:51 by woonchoi         ###   ########.fr       */
+/*   Updated: 2022/06/13 15:32:59 by jasong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
